c++: Optimize exact dynamic cast to final type by comparing the vptr
For final target classes of which the source type is a unique non-
virtual base, we know that the dynamic_cast succeeding is equivalent to
the vptr pointing to the target's vtable. So check this if possible
instead of calling the more expensive `__dynamic_cast` runtime method.

This optimization has been present in Clang since version 17 [1].

This might break[2] if the address of a particular vtable is not the
same throughout the program, i.e. when multiple shared libraries contain
it, but the linker does not canonicalize all references to the same
instance (because of hidden linkage or -Bsymbolic). For now, let's put
this behind an opt-in `-fassume-unique-vtables` flag.

[1] https://github.com/llvm/llvm-project/commit/9d525bf94b255df89587db955b5fa2d3c03c2c3e
[2] https://github.com/llvm/llvm-project/issues/120129

	PR/63164

gcc/c-family/ChangeLog:

	* c.opt: Add -fassume-unique-vtables

gcc/cp/ChangeLog:

	* rtti.cc (build_dynamic_cast_node): New helper function.
	(build_dynamic_cast_1): Compare vptr directly to target vtable
	address when the target is a final class type.

gcc/testsuite/ChangeLog:

	* g++.dg/rtti/dyncast-exact.C: New test.
	* g++.dg/rtti/dyncast-non-exact.C: New test.